cli: refactor `cloudcmd/validators`
